move all functions to lib.rs
- this matches the pattern in wgpu-rs
